# Vm string concatenation optimization