This pr optimizes the use-case where you are building up a string by repeatedly appending to the end. Instead of doing the concatenation immediately the vm now builds up a linked list of strings and only when the string is 'observed' will it perform the concatenation.

Depending on the sizes this can have a 2x - 100x speedup.

A simple benchmark that writes a list containing `100 000` integers to a file:
```
import "std.nov"
act main()
  l = rangeList(0, 100_000);
  benchReport(impure lambda ()
    fileWrite(Path("test.output.txt"), string(l))
  )
main()
```
Before:
```
Bench: 5s (5245334027 ns) (41s / 8 itrs)
Bench: 5s (5314192724 ns) (42s / 8 itrs)
Bench: 5s (5351752398 ns) (42s / 8 itrs)
```
After:
```
Bench: 46ms (46416539 ns) (371ms / 8 itrs)
Bench: 42ms (42643758 ns) (341ms / 8 itrs)
Bench: 42ms (42900834 ns) (343ms / 8 itrs)
```

A slightly more realistic example would be parsing and pretty printing a ~ 1 megabyte file.
```
import "std.nov"
act main()
  jp = jsonParser();
  jw = jsonWriter();
  benchReport(impure lambda ()
    p = Path("test.input.json");
    j = jp(fileRead(p));
    o = jw(j ?? toJsonVal(-1));
    fileWrite(Path("test.output.json"), o.string())
  )
main()
```
Before:
```
Bench: 11s (11429265968 ns) (1m31s / 8 itrs)
Bench: 10s (10929062345 ns) (1m27s / 8 itrs)
Bench: 11s (11283870238 ns) (1m30s / 8 itrs)
```
After:
```
Bench: 2s (2027739718 ns) (16s / 8 itrs)
Bench: 2s (2023707540 ns) (16s / 8 itrs)
Bench: 1s (1924621298 ns) (15s / 8 itrs)
```

At this time its only implemented for forward concatenation (appending to the end) but in theory should be implementable for the reverse case also.